Segregated game playing logic to the game package
