+++ tournament/tournament.go
@@ -1,12 +1,10 @@
 package tournament
 
 import (
-	"log"
 	"sync"
 
 	"github.com/gorilla/websocket"
 
-	"github.com/jar2333/MatchMaker/api"
 	"github.com/jar2333/MatchMaker/game"
 )
 
@@ -50,67 +48,14 @@ func (t *Tournament) Start() {
 
 func (t *Tournament) playGames() {
 	for g := range t.games {
-		go t.playGame(g)
+		conn1 := t.reg.GetConnection(g.P1())
+		conn2 := t.reg.GetConnection(g.P2())
+		go game.Play(g, conn1, conn2)
 	}
-}
-
-func (t *Tournament) playGame(g game.Game) {
-	// Get Registry
-	reg := &t.reg
-
-	// Get player keys for this game
-	p1 := g.P1()
-	p2 := g.P2()
-
-	// Grab a reference to the websockets corresponding to player 1 and player 2
-	conn1 := reg.GetConnection(p1)
-	conn2 := reg.GetConnection(p2)
-
-	// Game loop until game is finished and winner is found:
-	var msg []byte
-	var played bool
-	for !g.IsFinished() {
-		// Parse player 1's move, perform it, send game state
-		played = false
-		for !played {
-			msg = readMessage(conn1)
-			played = g.Play(p1, api.ParseMove(msg))
-		}
-		sendState(conn1, g)
-
-		if g.IsFinished() {
-			break
-		}
-
-		// Parse player 2's move, perform it, send game state
-		played = false
-		for !played {
-			msg = readMessage(conn2)
-			played = g.Play(p2, api.ParseMove(msg))
-		}
-		sendState(conn2, g)
-	}
-
-}
-
-func readMessage(conn *websocket.Conn) []byte {
-	for {
-		// Read message from websocket connection
-		_, msg, err := conn.ReadMessage()
-		if err != nil {
-			log.Println(err)
-			continue
-		}
-		return msg
-	}
-}
-
-func sendState(conn *websocket.Conn, g game.Game) {
-	// Not yet implemented
 }
 
 // =============================
-// = Match-making goroutines
+// = Tournament Match-making goroutines
 // =============================
 
 func (t *Tournament) matchMake() {
